requirements.txt
pypdf
pymysql
docx2txt
grandalf
pydantic
chromadb
streamlit
langchain
langgraph
SQLAlchemy
cryptography
python_dotenv
langchain-core
langchain-groq
langchain-ollama
typing_extensions
langchain-community
streamlit_option_menu
langchain-text-splitters
